Добавлен объект конфигурации с ключами, исправлена ошибка с bcrypt

+++ public/app.js
@@ -35,41 +35,45 @@ document.querySelectorAll('.product__price').forEach( productPrice => {
 
 const $cart = document.getElementById('cart');
 
-$cart.addEventListener('click', (e) => {
-    if(e.target.classList.contains('js-remove')) {
-       
-        const id = e.target.dataset.id;
-        const csrf = e.target.dataset.csrf;
+if($cart) {
+    $cart.addEventListener('click', (e) => {
+        if(e.target.classList.contains('js-remove')) {
+           
+            const id = e.target.dataset.id,
+                  csrf = e.target.dataset.csrf;
+    
+            fetch('cart/remove/' + id, {
+                method: 'delete',
+                headers: {
+                    'X-XSRF-TOKEN': csrf,
+                }
+            })
+            .then(res => res.json())
+            .then(cart => {
+                if(cart.products.length) {
+                    const html = cart.products.map( p => {
+                        const price = toCurrency(p.price);
+                        return `
+                        <tr>
+                            <td>${p.title}</td>
+                            <td>${p.quantity}</td>
+                            <td>${price}</td>
+                            <td>
+                                <button type="button" class="btn btn-small js-remove" data-id="${p.id}">Удалить</button>
+                            </td>
+                        </tr>
+                        `
+                    }).join('');
+                    $cart.querySelector('tbody').innerHTML = html;
+                    $cart.querySelector('.product__price').innerHTML = toCurrency(cart.price);
+                } else {
+                    $cart.innerHTML = '<p>Товаров в корзине пока нет</p>'
+                }
+            })
+    
+        }
+    })
+}
 
-        fetch('cart/remove/' + id, {
-            method: 'delete',
-            headers: {
-                'X-XSRF-TOKEN': csrf,
-            }
-        })
-        .then(res => res.json())
-        .then(cart => {
-            if(cart.products.length) {
-                const html = cart.products.map( p => {
-                    const price = toCurrency(p.price);
-                    return `
-                    <tr>
-                        <td>${p.title}</td>
-                        <td>${p.quantity}</td>
-                        <td>${price}</td>
-                        <td>
-                            <button type="button" class="btn btn-small js-remove" data-id="${p.id}">Удалить</button>
-                        </td>
-                    </tr>
-                    `
-                }).join('');
-                $cart.querySelector('tbody').innerHTML = html;
-                $cart.querySelector('.product__price').innerHTML = toCurrency(cart.price);
-            } else {
-                $cart.innerHTML = '<p>Товаров в корзине пока нет</p>'
-            }
-        })
 
-    }
-})
